fix(portal): TopNav の Score 表示を実データに繋ぐ
## 症状

Participant Portal の TopNavigation 右上に常時表示される widget が
「Score: — / Rank: —」固定 placeholder で、Home の累計スコアパネルが
3200 pt を表示していても TopNav には反映されない。

## 原因

`AppLayout.tsx:51-52` で `score` / `rank` が hardcoded `"—"`。
過去に「scoring backend (別 PR) が繋がるまで placeholder」と意図的に
保留したまま、scoring backend (Phase 2c で完成) との接続が漏れていた。

## 修正

新規 `auth/TeamViewProvider` を導入:

- `ShellLayout` 内で `/portal/me` を 5 秒 polling
- `useTeamView()` hook で view / error / refresh を context 共有
- TopNav は `view.problems[].score` の合計を表示
- Home page は自前の polling state を撤去して `useTeamView()` に統一
  (= polling が 1 つに集約、TopNav と Home が同じデータを共有)
- Rank は leaderboard API 接続前なので "—" 維持 (#163 follow-up で対応)

## 物理影響

| 種別 | リソース | 注 |
| --- | --- | --- |
| UPDATE | participant-portal SPA | TopNav score + Home polling 共通化 |
| NO-OP  | Lambda / DDB / IAM / CDK | 不変 |

## Test plan

- [x] make before-commit 緑 (248 件全 pass)
- [ ] AWS で実 deploy → backend mode で Portal にログイン → TopNav に実
      score (X pt) が表示されることを目視確認

## Regression 分析

- dev-mock モードでは backend を叩かないため TopNav score は "—" のまま
  (既存の Home behavior と整合)
- session 不在時は polling 起動しない (= /login や /setup で動かない)
- 全 problem が FAILED / DELETED に到達したら polling 停止 (Home と同じ条件)
- view 同一性チェック (`viewIsUnchanged`) を共通化、TopNav の no-op 再 render を
  Home と同じく抑制
